functions/graph/requirements.txt
google-cloud-storage==1.27.0
gcsfs==0.6.1
pandas==1.0.3
plotly==4.6.0
